requirements.txt
--extra-index-url https://pypi.cloud.soda.io soda-postgres
--extra-index-url https://pypi.cloud.soda.io soda-scientific
--extra-index-url https://pypi.cloud.soda.io soda-redshift
--extra-index-url https://pypi.cloud.soda.io soda-pandas-dask
agate==1.9.1
aiobotocore==2.13.2
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.4.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.33.33
Babel==2.15.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.8.0
boto3==1.34.151
botocore==1.34.162
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.4
colorama==0.4.6
coloredlogs==14.0
contourpy==1.2.1
croniter==3.0.3
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.10
daff==1.3.46
dagit==1.7.15
dagster==1.7.15
dagster-aws==0.23.15
dagster-cloud==1.7.15
dagster-cloud-cli==1.7.15
dagster-dbt==0.24.0
dagster-graphql==1.7.15
dagster-pipes==1.7.15
dagster-webserver==1.7.15
dask==2023.7.1
dask-sql==2023.8.0
db-dtypes==1.2.0
dbt==1.0.0.38.7
dbt-adapters==1.4.0
dbt-bigquery==1.8.2
dbt-common==1.7.0
dbt-core==1.8.4
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-redshift==1.8.1
dbt-semantic-interfaces==0.5.1
dbt-snowflake==1.8.3
dbt-trino==1.8.1
deepdiff==6.7.1
Deprecated==1.2.14
distributed==2023.7.1
dnspython==2.6.1
docstring_parser==0.16
docutils==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
github3.py==4.0.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.10.2
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-health-checking==1.62.2
grpcio-status==1.62.2
h11==0.14.0
holidays==0.53
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflect==7.3.1
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
kiwisolver==1.4.5
leather==0.4.0
locket==1.0.0
Logbook==1.5.3
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mashumaro==3.13.1
matplotlib==3.9.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==1.5.3
parsedatetime==2.6
partd==1.4.2
pathspec==0.12.1
pendulum==3.0.0
pex==2.13.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.36
prophet==1.1.5
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
redshift-connector==2.0.917
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.7.2
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3fs==2024.6.1
s3transfer==0.10.2
scipy==1.13.1
scramp==1.4.5
segment-analytics-python==2.3.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlglot==25.10.0
sqlglotrs==0.2.8
sqlparse==0.5.1
stanio==0.5.1
starlette==0.37.2
structlog==24.4.0
tabulate==0.9.0
tblib==3.0.0
text-unidecode==1.3
time-machine==2.14.2
tomli==2.0.1
tomlkit==0.13.0
toolz==0.12.1
toposort==1.10
tornado==6.4.1
tqdm==4.66.4
trino==0.329.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.3
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2
